Added search address by text
This is the first pass on figuring out what page address in the library is. I decided to use \t as delimiter because it is a control character, i.e. it will not be used by any implemented encode_strings, even the one with 128019 characters, printable unicode, is built without 'Cc' category. The only trouble will be if a user decides to implicitly specify \t in -c (custom charset flag) or -cf (custom charset file), and they are told not to.